REFACTOR: Updated The “loadHomeViewData” Action And HomeView.js
Problem
The loadHomeViewData action method did not append the payload data
to the new state object. Also, HomeView.js needs to be updated to reflect
the changes we made to the state.

Solution
The looadHomeViewData now captures the state properties within the payload.
Also, getUnpaidStudentData, getRecentStudents, getCurrentFoodsItems, and
setFakeTimeStamps functions are no longer being imported to format the
Hardcoded state data that was once used. Now useSelector is pulling the
appropriate data.